# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Payd Money Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payd/account:
    get:
      operationId: getAccount
      tags: [payd]
      summary: Get account info and balances
      responses:
        "200":
          description: Account and balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /payd/transactions:
    get:
      operationId: getTransactions
      tags: [payd]
      summary: Get transaction history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /payd/payin:
    post:
      operationId: initiatePayin
      tags: [payd]
      summary: Initiate a payin (deposit via M-Pesa STK push)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayinInput"
      responses:
        "200":
          description: Payin initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayinResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /payd/merchant:
    post:
      operationId: initiateMerchantPayout
      tags: [payd]
      summary: Pay to a Paybill or Till number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantPayoutInput"
      responses:
        "200":
          description: Merchant payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantPayoutResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /payd/p2p:
    post:
      operationId: initiateP2PTransfer
      tags: [payd]
      summary: Send money to a fellow Payd member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/P2PTransferInput"
      responses:
        "200":
          description: P2P transfer initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/P2PTransferResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /payd/tx-status/{reference}:
    get:
      operationId: getTransactionStatus
      tags: [payd]
      summary: Look up the status of a transaction by reference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionStatusResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /payd/payout:
    post:
      operationId: initiatePayout
      tags: [payd]
      summary: Initiate a payout (withdrawal to mobile money)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayoutInput"
      responses:
        "200":
          description: Payout initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /payd/summary:
    get:
      operationId: getSummary
      tags: [payd]
      summary: Get dashboard summary stats (total in, total out, recent counts)
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "500":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: ["string", "null"]
      required:
        - error
    Balance:
      type: object
      properties:
        currency:
          type: string
        available_balance:
          type: number
        ledger_balance:
          type: number
      required:
        - currency
        - available_balance
        - ledger_balance
    AccountInfo:
      type: object
      properties:
        username:
          type: string
        email:
          type: ["string", "null"]
        first_name:
          type: ["string", "null"]
        last_name:
          type: ["string", "null"]
        account_id:
          type: ["string", "null"]
        balances:
          type: array
          items:
            $ref: "#/components/schemas/Balance"
        connected:
          type: boolean
      required:
        - username
        - connected
        - balances
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
        narration:
          type: ["string", "null"]
        created_at:
          type: string
        reference:
          type: ["string", "null"]
        channel:
          type: ["string", "null"]
        phone_number:
          type: ["string", "null"]
      required:
        - id
        - type
        - amount
        - currency
        - status
        - created_at
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transactions
        - total
        - page
        - limit
    PayinInput:
      type: object
      properties:
        phone_number:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: KES
        channel:
          type: string
          default: MPESA
        narration:
          type: ["string", "null"]
      required:
        - phone_number
        - amount
    PayinResult:
      type: object
      properties:
        success:
          type: boolean
        reference:
          type: ["string", "null"]
        message:
          type: string
        transaction_reference:
          type: ["string", "null"]
      required:
        - success
        - message
    PayoutInput:
      type: object
      properties:
        phone_number:
          type: string
        amount:
          type: number
        currency:
          type: string
          default: KES
        network_code:
          type: string
          default: MPESA
        narration:
          type: ["string", "null"]
        account_id:
          type: ["string", "null"]
      required:
        - phone_number
        - amount
    PayoutResult:
      type: object
      properties:
        success:
          type: boolean
        reference:
          type: ["string", "null"]
        message:
          type: string
      required:
        - success
        - message
...
    MerchantPayoutInput:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
          default: KES
        phone_number:
          type: string
        narration:
          type: string
        business_type:
          type: string
          enum: [paybill, till]
          default: paybill
        business_account:
          type: string
        business_number:
          type: ["string", "null"]
        wallet_type:
          type: ["string", "null"]
      required:
        - amount
        - phone_number
        - narration
        - business_account
    MerchantPayoutResult:
      type: object
      properties:
        success:
          type: boolean
        correlator_id:
          type: ["string", "null"]
        message:
          type: string
        status:
          type: ["string", "null"]
      required:
        - success
        - message
    P2PTransferInput:
      type: object
      properties:
        receiver_username:
          type: string
        amount:
          type: number
        narration:
          type: string
        phone_number:
          type: string
        wallet_type:
          type: ["string", "null"]
      required:
        - receiver_username
        - amount
        - narration
        - phone_number
    P2PTransferResult:
      type: object
      properties:
        success:
          type: boolean
        transaction_reference:
          type: ["string", "null"]
        message:
          type: string
      required:
        - success
        - message
    TransactionStatusDetails:
      type: object
      properties:
        status:
          type: ["string", "null"]
        payer:
          type: ["string", "null"]
        receiver:
          type: ["string", "null"]
        phone_number:
          type: ["string", "null"]
        channel:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        merchant_id:
          type: ["string", "null"]
        account_number:
          type: ["string", "null"]
        email_address:
          type: ["string", "null"]
    TransactionStatusResult:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        currency:
          type: string
        amount:
          type: number
        balance:
          type: number
        type:
          type: string
        transaction_category:
          type: ["string", "null"]
        created_at:
          type: string
        transaction_details:
          $ref: "#/components/schemas/TransactionStatusDetails"
      required:
        - id
        - code
        - currency
        - amount
        - balance
        - type
        - created_at
    DashboardSummary:
      type: object
      properties:
        total_payin:
          type: number
        total_payout:
          type: number
        payin_count:
          type: integer
        payout_count:
          type: integer
        currency:
          type: string
        recent_activity_count:
          type: integer
      required:
        - total_payin
        - total_payout
        - payin_count
        - payout_count
        - currency
        - recent_activity_count